RF: Move format_utils to base
